Fix compilation of ERC20 example
